ms-novel-code/python-novel-code/host_tasks/task_colab-122/requirements.txt
cupy-cuda12x==13.4.1
numpy==2.2.6
